Moved tests of regrid module from backend/tests to tests
