--- backend/tests/unit/__init__.py
@@ -1,8 +1,0 @@
-"""
-Unit tests for the :mod:`esmvaltool.backend` package.
-
-"""
-
-from __future__ import (absolute_import, division, print_function)
-from six.moves import (filter, input, map, range, zip)  # noqa
-import six  # noqa
