+++ tests/unit/regrid/test__stock_cube.py
